change back to mp4 audio encode
Because flutter_sound dependency doesn't support FLAC encode on Android, change it back and just go with openai api.

The implementation of making use of Google stt + groq llama3 gonna be taken for normal user (free using), and that should be record audio by m4a for iOS and amr wb for Android.